really add dependence template support.

+++ nixify.sh
@@ -38,22 +38,23 @@ case $3 in
 
          "${SCRIPT_ABS_PATH}"/fix-up/do.sh "${quick_cook_nixify_tmp_dir}" "$2"
 
-         for DIR_NAME1 in "ci" "cd" "nix" "config"
+         for DIR_NAME1 in "ci" "cd" "nix" "env" "config" "k8s"
          do
              [ -d "$1/$2/$DIR_NAME1" ] && mv "$1/$2/$DIR_NAME1" "$1/$2/$DIR_NAME1.bak.by.nixify.$(date +%Y%m%d%H%M%S)"
-             cp -R "${quick_cook_nixify_tmp_dir}/$2/$DIR_NAME1" "$1/$2"
+             [ -d  "${quick_cook_nixify_tmp_dir}/$2/$DIR_NAME1" ] && cp -R "${quick_cook_nixify_tmp_dir}/$2/$DIR_NAME1" "$1/$2"
          done
 
          for FILE_NAME1 in "default.nix" "shell.nix" "release.nix" "cross-build.nix" \
-             "docker.nix" "tarball.nix" "arion-pkgs.nix" "arion-compose.nix" \
+             "docker.nix" "tarball.nix" ".editorconfig" ".env" \
              "hie.yaml" "fourmolu.yaml" ".gitignore" ".dir-locals.el" ".ghci" \
-             ".hlint.yaml" "develop" "build" "deploy" "arion"
+             ".hlint.yaml" "develop" "build" "deploy" ".ghcid" ".ghcid.lib" ".ghcid.app"
          do
              [ -f "$1/$2/$FILE_NAME1" ] && mv "$1/$2/$FILE_NAME1" "$1/$2/$FILE_NAME1.nix.bak.by.nixify.$(date +%Y%m%d%H%M%S)"
-             cp "${quick_cook_nixify_tmp_dir}/$2/$FILE_NAME1" "$1/$2"
+             [ -f "${quick_cook_nixify_tmp_dir}/$2/$FILE_NAME1" ] && cp "${quick_cook_nixify_tmp_dir}/$2/$FILE_NAME1" "$1/$2"
          done
 
-         [ ! -f "$1/$2/cabal.project.local" ] && cp "${quick_cook_nixify_tmp_dir}/$2/cabal.project.local" "$1/$2"
+         [ ! -f "$1/$2/cabal.project.local" ] && [ -f "${quick_cook_nixify_tmp_dir}/$2/cabal.project.local" ] \
+             && cp "${quick_cook_nixify_tmp_dir}/$2/cabal.project.local" "$1/$2"
          if [ -f "$1/$2/cabal.project" ]; then
              THE_INDEX_STATE=$(grep index-state "${quick_cook_nixify_tmp_dir}/$2/cabal.project")
              grep -v "index-state" "$1/$2/cabal.project" > "${quick_cook_nixify_tmp_dir}/$2/cabal.project.orig"
@@ -61,15 +62,13 @@ case $3 in
              mv "$1/$2/cabal.project" "$1/$2/cabal.project.bak.by.nixify.$(date +%Y%m%d%H%M%S)"
              cp "${quick_cook_nixify_tmp_dir}/$2/cabal.project.orig" "$1/$2/cabal.project"
          else
-             cp "${quick_cook_nixify_tmp_dir}/$2/cabal.project" "$1/$2"
+             [ -f "${quick_cook_nixify_tmp_dir}/$2/cabal.project" ] && cp "${quick_cook_nixify_tmp_dir}/$2/cabal.project" "$1/$2"
          fi
 
-	     # for executables, we need to copy to override them after applying tempalte(Why?)
-	     cp "${SCRIPT_ABS_PATH}"/deployment-framework/arion "$1/$2/cd/" && chmod +x "$1/$2/cd/arion"
          mkdir -p "$1/$2/.github/workflows"
          [ -f "$1/$2/.github/workflows/nix-ci.yml" ] \
              && mv "$1/$2/.github/workflows/nix-ci.yml" "$1/$2/.github/workflows/nix-ci.yml.bak.by.nixify.$(date +%Y%m%d%H%M%S)"
-         cp "${quick_cook_nixify_tmp_dir}/$2/.github/workflows/nix-ci.yml" "$1/$2/.github/workflows/"
+         [ -f "${quick_cook_nixify_tmp_dir}/$2/.github/workflows/nix-ci.yml" ] && cp "${quick_cook_nixify_tmp_dir}/$2/.github/workflows/nix-ci.yml" "$1/$2/.github/workflows/"
 
          if [ -f "$1/$2/$2.cabal" ] && [ -f "$1/$2/stack.yaml" ]; then
              mv "$1/$2/stack.yaml" "$1/$2/stack.yaml.bak.by.nixify.$(date +%Y%m%d%H%M%S)"
